erdjs does not depend on "axios" and "fs" anymore. Changed the way an AbiRegistry is created.
