+++ src/smartcontracts/code.ts
@@ -1,5 +1,3 @@
-import axios, { AxiosResponse } from "axios";
-
 /**
  * Bytecode of a Smart Contract, as an abstraction.
  */
@@ -15,22 +13,6 @@ export class Code {
      */
     static fromBuffer(code: Buffer): Code {
         return new Code(code.toString("hex"));
-    }
-
-    /**
-     * Creates a Code object by loading the bytecode from a specified URL (WASM file).
-     */
-    static async fromUrl(url: string): Promise<Code> {
-        let response: AxiosResponse<ArrayBuffer> = await axios.get(url, {
-            responseType: 'arraybuffer',
-            transformResponse: [],
-            headers: {
-                "Accept": "application/wasm"
-            }
-        });
-
-        let buffer = Buffer.from(response.data);
-        return Code.fromBuffer(buffer);
     }
 
     /**
