+++ src/smartcontracts/interaction.spec.ts
@@ -109,7 +109,7 @@ describe("test smart contract interactor", function() {
     it("should interact with 'answer'", async function () {
         setupUnitTestWatcherTimeouts();
 
-        let abiRegistry = await loadAbiRegistry(["src/testdata/answer.abi.json"]);
+        let abiRegistry = await loadAbiRegistry("src/testdata/answer.abi.json");
         let abi = new SmartContractAbi(abiRegistry, ["answer"]);
         let contract = new SmartContract({ address: dummyAddress, abi: abi });
         let controller = new ContractController(provider);
@@ -171,7 +171,7 @@ describe("test smart contract interactor", function() {
     it("should interact with 'counter'", async function() {
         setupUnitTestWatcherTimeouts();
 
-        let abiRegistry = await loadAbiRegistry(["src/testdata/counter.abi.json"]);
+        let abiRegistry = await loadAbiRegistry("src/testdata/counter.abi.json");
         let abi = new SmartContractAbi(abiRegistry, ["counter"]);
         let contract = new SmartContract({ address: dummyAddress, abi: abi });
         let controller = new ContractController(provider);
@@ -218,7 +218,7 @@ describe("test smart contract interactor", function() {
     it("should interact with 'lottery-esdt'", async function() {
         setupUnitTestWatcherTimeouts();
 
-        let abiRegistry = await loadAbiRegistry(["src/testdata/lottery-esdt.abi.json"]);
+        let abiRegistry = await loadAbiRegistry("src/testdata/lottery-esdt.abi.json");
         let abi = new SmartContractAbi(abiRegistry, ["Lottery"]);
         let contract = new SmartContract({ address: dummyAddress, abi: abi });
         let controller = new ContractController(provider);
