+++ src/smartcontracts/typesystem/abiRegistry.spec.ts
@@ -1,7 +1,6 @@
 import { assert } from "chai";
-import { extendAbiRegistry, loadAbiRegistry } from "../../testutils";
+import { loadAbiRegistry } from "../../testutils";
 import { BinaryCodec } from "../codec";
-import { AbiRegistry } from "./abiRegistry";
 import { AddressType } from "./address";
 import { OptionalType } from "./algebraic";
 import { BytesType } from "./bytes";
@@ -12,46 +11,21 @@ import { StructType } from "./struct";
 import { TokenIdentifierType } from "./tokenIdentifier";
 
 describe("test abi registry", () => {
-    it("should extend", async () => {
-        let registry = new AbiRegistry();
-
-        await extendAbiRegistry(registry, "src/testdata/answer.abi.json");
-        assert.lengthOf(registry.interfaces, 1);
-        assert.lengthOf(registry.customTypes, 0);
-        assert.lengthOf(registry.getInterface("answer").endpoints, 1);
-
-        await extendAbiRegistry(registry, "src/testdata/counter.abi.json");
-        assert.lengthOf(registry.interfaces, 2);
-        assert.lengthOf(registry.customTypes, 0);
-        assert.lengthOf(registry.getInterface("counter").endpoints, 3);
-
-        await extendAbiRegistry(registry, "src/testdata/lottery-esdt.abi.json");
-        assert.lengthOf(registry.interfaces, 3);
-        assert.lengthOf(registry.customTypes, 2);
-
-        assert.lengthOf(registry.getInterface("Lottery").endpoints, 7);
-        assert.lengthOf(registry.getStruct("LotteryInfo").getFieldsDefinitions(), 7);
-        assert.lengthOf(registry.getEnum("Status").variants, 3);
-    });
-
     it("load should also remap known to types", async () => {
-        let registry = await loadAbiRegistry([
-            "src/testdata/answer.abi.json",
-            "src/testdata/counter.abi.json",
-            "src/testdata/lottery-esdt.abi.json",
-        ]);
-
         // Ultimate answer
+        let registry = await loadAbiRegistry("src/testdata/answer.abi.json");
         let answer = registry.getInterface("answer");
         let getUltimateAnswer = answer.getEndpoint("getUltimateAnswer");
         assert.instanceOf(getUltimateAnswer.output[0].type, I64Type);
 
         // Counter
+        registry = await loadAbiRegistry("src/testdata/counter.abi.json");
         let counter = registry.getInterface("counter");
         let getCounter = counter.getEndpoint("get");
         assert.instanceOf(getCounter.output[0].type, I64Type);
 
         // Lottery
+        registry = await loadAbiRegistry("src/testdata/lottery-esdt.abi.json");
         let lottery = registry.getInterface("Lottery");
         let start = lottery.getEndpoint("start");
         let getStatus = lottery.getEndpoint("status");
@@ -90,7 +64,7 @@ describe("test abi registry", () => {
             "hex"
         );
 
-        let registry = await loadAbiRegistry(["src/testdata/multisig.abi.json"]);
+        let registry = await loadAbiRegistry("src/testdata/multisig.abi.json");
         let multisig = registry.getInterface("Multisig");
         let performAction = multisig.getEndpoint("getActionData");
         assert.equal(performAction.output[0].type.getName(), "Action");
