+++ src/testutils/utils.ts
@@ -7,6 +7,7 @@ import { Transaction } from "../transaction";
 import { TransactionWatcher } from "../transactionWatcher";
 import { IChainID, IGasLimit } from "../interface";
 import { TestWallet } from "./wallets";
+import axios, { AxiosResponse } from "axios";
 
 export async function prepareDeployment(obj: {
     deployer: TestWallet,
@@ -37,32 +38,33 @@ export async function prepareDeployment(obj: {
 
 export async function loadContractCode(path: PathLike): Promise<Code> {
     if (isOnBrowserTests()) {
-        return Code.fromUrl(path.toString());
-    }
+        let response: AxiosResponse<ArrayBuffer> = await axios.get(path.toString(), {
+            responseType: 'arraybuffer',
+            transformResponse: [],
+            headers: {
+                "Accept": "application/wasm"
+            }
+        });
 
+        let buffer = Buffer.from(response.data);
+        return Code.fromBuffer(buffer);
+    }
+    
     // Load from file.
     let buffer: Buffer = await fs.promises.readFile(path);
     return Code.fromBuffer(buffer);
 }
 
-export async function loadAbiRegistry(paths: PathLike[]): Promise<AbiRegistry> {
-    let sources = paths.map(e => e.toString());
-
+export async function loadAbiRegistry(path: PathLike): Promise<AbiRegistry> {
     if (isOnBrowserTests()) {
-        return AbiRegistry.load({ urls: sources });
+        let response: AxiosResponse = await axios.get(path.toString());
+        return AbiRegistry.create(response.data);
     }
 
-    return AbiRegistry.load({ files: sources });
-}
-
-export async function extendAbiRegistry(registry: AbiRegistry, path: PathLike): Promise<AbiRegistry> {
-    let source = path.toString();
-
-    if (isOnBrowserTests()) {
-        return registry.extendFromUrl(source);
-    }
-
-    return registry.extendFromFile(source);
+    // Load from files
+    let jsonContent: string = await fs.promises.readFile(path, { encoding: "utf8" });
+    let json = JSON.parse(jsonContent);
+    return AbiRegistry.create(json);
 }
 
 export function isOnBrowserTests() {
